Simplify API layer
