+++ api/exchange_apis/api_protocol.py
@@ -3,13 +3,9 @@ from typing import Protocol, Any
 
 class ExchangeApiProtocol(Protocol):
     """
-    Protocol for exchange API classes (BinanceApi, KucoinApi, etc).
-    Add all methods and attributes you want to call in a type-safe way.
-
-    This is mainly for mypy type checking and IDE autocompletion.
+    Minimal shared protocol used generically by controllers across exchanges.
     """
 
-    # --- Common Exchange Methods ---
     def get_server_time(self) -> Any: ...
     def get_order(self, order_id: str) -> dict: ...
     def add_order(
@@ -22,78 +18,10 @@ class ExchangeApiProtocol(Protocol):
         time_in_force: Any = ...,
     ) -> Any: ...
     def cancel_order(self, order_id: str) -> dict: ...
-    def cancel_all_orders(self, symbol: str = None) -> dict: ...
-    def get_open_orders(self, symbol: str = None) -> Any: ...
+    def cancel_all_orders(self, symbol: str | None = None) -> dict: ...
+    def get_open_orders(self, symbol: str | None = None) -> Any: ...
     def get_all_symbols(self) -> Any: ...
     def get_part_order_book(self, symbol: str, size: str) -> Any: ...
-
-    # --- Binance-specific Methods ---
-    def request(
-        self,
-        url: str,
-        method: str = "GET",
-        session: Any = None,
-        payload: dict = None,
-        **kwargs,
-    ) -> Any: ...
-    def signed_request(
-        self, url: str, method: str = "GET", payload: dict = {}
-    ) -> dict: ...
-    def get_listen_key(self) -> str: ...
-    def exchange_info(self, symbol: str = None) -> Any: ...
-    def price_filter_by_symbol(self, symbol: str, filter_limit: str) -> Any: ...
-    def lot_size_by_symbol(self, symbol: str, lot_size_limit: str) -> Any: ...
-    def min_notional_by_symbol(
-        self, symbol: str, min_notional_limit: str = "minNotional"
-    ) -> Any: ...
-    def _calculate_price_precision(self, symbol: str) -> int: ...
-    def _calculate_qty_precision(self, symbol: str) -> int: ...
-    def ticker_24(self, type: str = "FULL", symbol: str = None) -> Any: ...
-    def get_ticker_price(self, symbol: str) -> Any: ...
-    def last_ticker_price(self, symbol: str) -> float: ...
-    def get_raw_klines(self, *args, **kwargs) -> Any: ...
-    def get_account_balance(self) -> Any: ...
-    def get_wallet_balance(self) -> Any: ...
-    def buy_isolated_margin_order(self, *args, **kwargs) -> Any: ...
-    def sell_isolated_margin_order(self, *args, **kwargs) -> Any: ...
-    def cancel_margin_order(self, symbol: str, order_id: int) -> Any: ...
-    def enable_isolated_margin_account(self, symbol: str) -> Any: ...
-    def disable_isolated_margin_account(self, symbol: str) -> Any: ...
-    def get_isolated_account(self, symbol: str) -> Any: ...
-    def transfer_isolated_margin_to_spot(
-        self, asset: str, symbol: str, amount: float
-    ) -> Any: ...
-    def transfer_spot_to_isolated_margin(
-        self, asset: str, symbol: str, amount: float
-    ) -> Any: ...
-    def create_margin_loan(
-        self, asset: str, symbol: str, amount: float, isIsolated: bool = True
-    ) -> Any: ...
-    def get_max_borrow(self, asset: str, isolated_symbol: str = None) -> Any: ...
-    def get_margin_loan_details(self, loan_id: int, symbol: str) -> Any: ...
-    def get_repay_details(self, loan_id: int, symbol: str) -> Any: ...
-    def repay_margin_loan(self, *args, **kwargs) -> Any: ...
-    def manual_liquidation(self, symbol: str) -> Any: ...
-    def get_interest_history(self, asset: str, symbol: str) -> Any: ...
-    def get_isolated_balance(self, symbol: str = None) -> list: ...
-    def get_isolated_balance_total(self) -> Any: ...
-    def transfer_dust(self, assets: list[str]) -> Any: ...
-    def query_open_orders(self, symbol: str) -> Any: ...
-    def get_all_orders(
-        self, symbol: str, order_id: int = 0, start_time: Any = None
-    ) -> Any: ...
-    def post_sell_order(self, *args, **kwargs) -> Any: ...
-    def post_buy_order(self, *args, **kwargs) -> Any: ...
-    def delete_opened_order(self, symbol: str, order_id: Any) -> Any: ...
-    def delete_order(self, symbol: str, order_id: int) -> Any: ...
-    def close_all_orders(self, symbol: str) -> Any: ...
-    def get_book_depth(self, symbol: str) -> dict: ...
-    def get_user_asset(self, asset: str, need_btc_valuation: bool = False) -> Any: ...
-    def ticker_24_pct_change(self, *args, **kwargs) -> Any: ...
-    def ticker_24_last_price_cached(self, *args, **kwargs) -> Any: ...
-    def get_tags(self, symbol: str) -> dict: ...
-
-    # --- Kucoin-specific Methods ---
     def add_margin_order(
         self,
         symbol: str,
