+++ api/orders/routes.py
@@ -3,7 +3,8 @@ from tools.exceptions import BinanceErrors
 from tools.handle_error import json_response, json_response_error
 from orders.controller import OrderFactory
 from orders.schemas import OrderParams
-
+from exchange_apis.binance.orders import BinanceOrderController
+from exchange_apis.binance.base import BinanceApi
 
 order_blueprint = APIRouter()
 
@@ -23,8 +24,7 @@ def create_sell_order(item: OrderParams):
 @order_blueprint.delete("/close/{symbol}/{orderid}", tags=["orders"])
 def delete_order(symbol, orderid):
     try:
-        order = OrderFactory().get_order_controller()
-        data = order.delete_order(symbol, orderid)
+        data = BinanceApi().delete_order(symbol, orderid)
         resp = json_response({"message": "Order deleted!", "data": data})
     except BinanceErrors as error:
         resp = json_response_error(error.message)
@@ -35,17 +35,8 @@ def delete_order(symbol, orderid):
 
 @order_blueprint.get("/margin/sell/{symbol}/{qty}", tags=["orders"])
 def margin_sell(symbol, qty):
-    order = OrderFactory().get_order_controller()
-    return order.sell_margin_order(symbol, qty)
-
-
-@order_blueprint.get("/all-orders", tags=["orders"])
-def get_all_orders(symbol, order_id=0, start_time=None):
     try:
-        order = OrderFactory().get_order_controller()
-        data = order.get_all_orders(symbol, order_id=order_id, start_time=start_time)
-        return json_response({"message": "Orders found!", "data": data})
-    except ValueError as error:
-        return json_response_error(error.args[0])
+        data = BinanceOrderController().sell_margin_order(symbol, qty)
+        return json_response({"message": "Margin Sell order placed", "data": data})
     except BinanceErrors as error:
         return json_response_error(error.message)
