+++ api/streaming/streaming_controller.py
@@ -13,7 +13,7 @@ from databases.crud.symbols_crud import SymbolsCrud
 from deals.abstractions.factory import DealAbstract
 from deals.margin import MarginDeal
 from deals.spot import SpotLongDeal
-from exchange_apis.binance import BinanceApi
+from exchange_apis.binance.base import BinanceApi
 from streaming.models import HABollinguerSpread
 from tools.enum_definitions import Status, Strategy
 from tools.exceptions import BinanceErrors, BinbotErrors
